requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
python-dotenv==1.0.0
aiohttp
cryptography
pyjwt
telegram
flask
websockets
httpx
openai
